+++ test/integration/tests/ghc-install-hooks/Main.hs
@@ -1,0 +1,7 @@
+import System.Process (rawSystem)
+import Control.Exception (throwIO)
+import StackTest
+import Control.Monad (unless)
+
+main :: IO ()
+main = unless isWindows $ rawSystem "bash" ["run.sh"] >>= throwIO